MAYA-128473 - Move UFE to its own Project - Part 10
* Functions moved to UsdUfe:
** removeDescendants / recreateDescendants
** convertLoadRulesToText / setLoadRulesFromText / createLoadRulesFromText
* Commands moved to UsdUfe:
** UsdUndoLoadPayloadCommand (+ python binding LoadPayloadCommand)
** UsdUndoUnloadPayloadCommand (+ binding UnloadPayloadCommand)
** SetVariantSelectionCommand
* Classes moved to UsdUfe:
** UsdContextOpsHandler/UsdContextOps. Split this class into Usd
   and Maya versions.
* New optional DCC Function: saveStageLoadRules
