+++ lib/mayaUsd/python/CMakeLists.txt
@@ -30,7 +30,6 @@ target_sources(${PYTHON_TARGET_NAME}
         wrapAdaptor.cpp
         wrapBlockSceneModificationContext.cpp
         wrapColorSpace.cpp
-        wrapCommands.cpp
         wrapConverter.cpp
         wrapDiagnosticDelegate.cpp
         wrapMeshWriteUtils.cpp
