+++ lib/mayaUsd/python/module.cpp
@@ -27,7 +27,6 @@ TF_WRAP_MODULE
     TF_WRAP(Adaptor);
     TF_WRAP(BlockSceneModificationContext);
     TF_WRAP(ColorSpace);
-    TF_WRAP(Commands);
     TF_WRAP(Converter);
     TF_WRAP(ConverterArgs);
     TF_WRAP(DiagnosticDelegate);
