+++ lib/mayaUsd/ufe/ProxyShapeContextOpsHandler.cpp
@@ -15,7 +15,7 @@
 //
 #include "ProxyShapeContextOpsHandler.h"
 
-#include <mayaUsd/ufe/UsdContextOpsHandler.h>
+#include <mayaUsd/ufe/MayaUsdContextOpsHandler.h>
 #include <mayaUsd/ufe/Utils.h>
 
 #include <usdUfe/ufe/UsdSceneItem.h>
@@ -53,7 +53,7 @@ Ufe::ContextOps::Ptr ProxyShapeContextOpsHandler::contextOps(const Ufe::SceneIte
         PXR_NS::UsdStageWeakPtr stage = getStage(item->path());
         if (stage) {
             auto               usdItem = UsdSceneItem::create(item->path(), stage->GetPseudoRoot());
-            auto               usdContextOpsHandler = UsdContextOpsHandler::create();
+            auto               usdContextOpsHandler = MayaUsdContextOpsHandler::create();
             auto               cOps = usdContextOpsHandler->contextOps(usdItem);
             UsdContextOps::Ptr usdCOps = std::dynamic_pointer_cast<UsdContextOps>(cOps);
