+++ lib/mayaUsd/utils/loadRules.h
@@ -20,31 +20,10 @@
 #include <mayaUsd/nodes/proxyShapeBase.h>
 
 #include <pxr/usd/usd/stage.h>
-#include <pxr/usd/usd/stageLoadRules.h>
 
 #include <maya/MApiNamespace.h>
 
 namespace MAYAUSD_NS_DEF {
-
-/*! \brief convert the stage load rules to a text format.
- */
-MAYAUSD_CORE_PUBLIC
-MString convertLoadRulesToText(const PXR_NS::UsdStage& stage);
-
-/*! \brief set the stage load rules from a text format.
- */
-MAYAUSD_CORE_PUBLIC
-void setLoadRulesFromText(PXR_NS::UsdStage& stage, const MString& text);
-
-/*! \brief convert the load rules to a text format.
- */
-MAYAUSD_CORE_PUBLIC
-MString convertLoadRulesToText(const PXR_NS::UsdStageLoadRules& rules);
-
-/*! \brief create load rules from a text format.
- */
-MAYAUSD_CORE_PUBLIC
-PXR_NS::UsdStageLoadRules createLoadRulesFromText(const MString& text);
 
 /*! \brief verify if there is a dynamic attribute on the object for load rules.
  */
