+++ lib/usdUfe/ufe/UsdContextOps.h
@@ -15,8 +15,7 @@
 //
 #pragma once
 
-#include <mayaUsd/base/api.h>
-
+#include <usdUfe/base/api.h>
 #include <usdUfe/ufe/UsdSceneItem.h>
 
 #include <pxr/usd/usd/prim.h>
@@ -24,8 +23,9 @@
 #include <ufe/contextOps.h>
 #include <ufe/path.h>
 
-namespace MAYAUSD_NS_DEF {
-namespace ufe {
+#include <vector>
+
+namespace USDUFE_NS_DEF {
 
 struct SchemaTypeGroup
 {
@@ -43,7 +43,7 @@ struct SchemaTypeGroup
 
     \see UFE ContextOps class documentation for more details
 */
-class MAYAUSD_CORE_PUBLIC UsdContextOps : public Ufe::ContextOps
+class USDUFE_PUBLIC UsdContextOps : public Ufe::ContextOps
 {
 public:
     typedef std::shared_ptr<UsdContextOps> Ptr;
@@ -71,8 +71,8 @@ public:
             return PXR_NS::UsdPrim();
     }
 
-    // When we are created from the ProxyShapeContextOpsHandler we do not have the proper
-    // Maya UFE scene item. So it won't return the correct node type. Therefore we set
+    // When we are created from a gateway node ContextOpsHandler we do not have the proper
+    // UFE scene item. So it won't return the correct node type. Therefore we set
     // this flag directly.
     void setIsAGatewayType(bool t) { fIsAGatewayType = t; }
     bool isAGatewayType() const { return fIsAGatewayType; }
@@ -82,7 +82,7 @@ public:
     Items                     getItems(const ItemPath& itemPath) const override;
     Ufe::UndoableCommand::Ptr doOpCmd(const ItemPath& itemPath) override;
 
-private:
+protected:
     UsdSceneItem::Ptr fItem;
     bool              fIsAGatewayType { false };
 
@@ -91,5 +91,4 @@ private:
 
 }; // UsdContextOps
 
-} // namespace ufe
-} // namespace MAYAUSD_NS_DEF
+} // namespace USDUFE_NS_DEF
