+++ lib/usdUfe/ufe/UsdContextOpsHandler.cpp
@@ -1,0 +1,46 @@
+//
+// Copyright 2020 Autodesk
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+#include "UsdContextOpsHandler.h"
+
+namespace USDUFE_NS_DEF {
+
+UsdContextOpsHandler::UsdContextOpsHandler()
+    : Ufe::ContextOpsHandler()
+{
+}
+
+UsdContextOpsHandler::~UsdContextOpsHandler() { }
+
+/*static*/
+UsdContextOpsHandler::Ptr UsdContextOpsHandler::create()
+{
+    return std::make_shared<UsdContextOpsHandler>();
+}
+
+//------------------------------------------------------------------------------
+// Ufe::ContextOpsHandler overrides
+//------------------------------------------------------------------------------
+
+Ufe::ContextOps::Ptr UsdContextOpsHandler::contextOps(const Ufe::SceneItem::Ptr& item) const
+{
+    UsdSceneItem::Ptr usdItem = std::dynamic_pointer_cast<UsdSceneItem>(item);
+#if !defined(NDEBUG)
+    assert(usdItem);
+#endif
+    return UsdContextOps::create(usdItem);
+}
+
+} // namespace USDUFE_NS_DEF
